[FEAT] 피드 댓글, 피드, 피드 좋아요 권한 관련 예외 처리 : 피드 작성 본인만 관련 기능들 수행할 수 있게 처리.

+++ src/main/java/com/example/trip/service/FeedService.java
@@ -1,5 +1,8 @@
 package com.example.trip.service;
 
+import com.example.trip.advice.exception.AuthFeedCommentNotFoundException;
+import com.example.trip.advice.exception.AuthFeedNotFoundException;
+import com.example.trip.advice.exception.AuthLikesNotFoundException;
 import com.example.trip.domain.*;
 import com.example.trip.dto.request.FeedRequestDto;
 import com.example.trip.repository.LikeRepository;
@@ -24,14 +27,15 @@ public class FeedService {
     private final LikeRepository likeRepository;
     private final BookMarkRepository bookMarkRepository;
     private final CommentRepository commentRepository;
-private final AwsS3Service awsS3Service;
-    public List<Feed> findAll(){
+    private final AwsS3Service awsS3Service;
+
+    public List<Feed> findAll() {
 
         return feedRepository.findAll();
     }
 
     @Transactional
-    public List<Long> registerFeed(User user, FeedRequestDto.FeedRequestRegisterDto feedRequestRegisterDto){
+    public List<Long> registerFeed(User user, FeedRequestDto.FeedRequestRegisterDto feedRequestRegisterDto) {
         // feed 저장
         Feed feed = Feed.builder()
                 .user(user)
@@ -51,7 +55,7 @@ private final AwsS3Service awsS3Service;
 
         List<List<FeedDetailLoc>> feedDetailLocs = feedDetails.stream()
                 .map(FeedDetail::getFeedDetailLoc)
-                        .collect(Collectors.toList());
+                .collect(Collectors.toList());
 
         feedDetailLocs.stream()
                 .forEach(x -> feedDetailLocRepository.saveAll(x));
@@ -73,8 +77,8 @@ private final AwsS3Service awsS3Service;
         return newFeedDetailLocs;
     }
 
-    public Map<String, String> registerFeedImage(Long feedDetailLocId, List<MultipartFile> imgFiles){
-        Map<String, String> nameAndUrl  = awsS3Service.uploadFile(imgFiles);
+    public Map<String, String> registerFeedImage(Long feedDetailLocId, List<MultipartFile> imgFiles) {
+        Map<String, String> nameAndUrl = awsS3Service.uploadFile(imgFiles);
         FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException(""));
         nameAndUrl.entrySet().forEach(x -> feedDetailLocImgRepository.save(FeedDetailLocImg.builder()
                 .feedDetailLoc(feedDetailLoc)
@@ -85,13 +89,22 @@ private final AwsS3Service awsS3Service;
         return nameAndUrl;
     }
 
-    public void deleteFeedImage(FeedRequestDto.FeedRequestDeleteImgDto  FeedRequestDeleteImgDto){
-        List<String> fileNames  = FeedRequestDeleteImgDto.getFileNames();
+    public void deleteFeedImage(FeedRequestDto.FeedRequestDeleteImgDto FeedRequestDeleteImgDto) {
+
+        List<String> fileNames = FeedRequestDeleteImgDto.getFileNames();
         fileNames.forEach(x -> awsS3Service.deleteFile(x));
     }
 
     @Transactional
-    public void modifyFeed(Long feedId, FeedRequestDto.FeedRequestModifyDto feedRequestModifyDto){
+    public void modifyFeed(User user, Long feedId, FeedRequestDto.FeedRequestModifyDto feedRequestModifyDto) {
+        // 피드를 올린 사람만 권한이 있어야함
+        List<Feed> myFeed = feedRepository.findByIdAndUserId(feedId, user.getId());
+
+        if (myFeed.isEmpty()) {
+            throw new AuthFeedNotFoundException();
+        }
+
+
         Feed feed = feedRepository.findById(feedId).orElseThrow(() -> new IllegalArgumentException("ID가 존재하지 않습니다. "));
 
         //feed Detail 수정
@@ -109,11 +122,17 @@ private final AwsS3Service awsS3Service;
         feed.update(feedRequestModifyDto);
     }
 
-    public void deleteFeed(Long feedId){
+    public void deleteFeed(User user, Long feedId) {
+        // 피드를 올린 사람만 권한이 있어야함
+        List<Feed> myFeed = feedRepository.findByIdAndUserId(feedId, user.getId());
+
+        if (myFeed.isEmpty()) {
+            throw new AuthFeedNotFoundException();
+        }
         feedRepository.deleteById(feedId);
     }
 
-    public void likeFeed(Long feedDetailLocId, User user){
+    public void likeFeed(Long feedDetailLocId, User user) {
         FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
         Likes like = Likes.builder()
                 .feedDetailLoc(feedDetailLoc)
@@ -124,11 +143,17 @@ private final AwsS3Service awsS3Service;
     }
 
 
-    public void unlikeFeed(Long feedId, User user){
-        likeRepository.deleteLikeFeed(feedId, user.getId());
+    public void unlikeFeed(Long feedDetailLocId, User user) {
+        // 좋아요를 한 사람만 권한이 있어야함
+        List<Likes> myLike = likeRepository.findByFeedDetailLocIdAndUserId(feedDetailLocId, user.getId());
+
+        if (myLike.isEmpty()) {
+            throw new AuthLikesNotFoundException();
+        }
+        likeRepository.deleteLikeFeed(feedDetailLocId, user.getId());
     }
 
-    public void bookmarkFeed(Long feedDetailLocId, User user){
+    public void bookmarkFeed(Long feedDetailLocId, User user) {
         FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
         BookMark bookmark = BookMark.builder()
                 .feedDetailLoc(feedDetailLoc)
@@ -137,11 +162,12 @@ private final AwsS3Service awsS3Service;
 
         bookMarkRepository.save(bookmark);
     }
-    public void unbookmarkFeed(Long feedDetailLocId, User user){
+
+    public void unbookmarkFeed(Long feedDetailLocId, User user) {
         bookMarkRepository.deleteBookmarkFeed(feedDetailLocId, user.getId());
     }
 
-    public void registerFeedComment(User user, Long feedDetailLocId, FeedRequestDto.FeedRequestCommentRegisterDto feedRequestCommentRegisterDto){
+    public void registerFeedComment(User user, Long feedDetailLocId, FeedRequestDto.FeedRequestCommentRegisterDto feedRequestCommentRegisterDto) {
         FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
         FeedComment feedComment = FeedComment.builder()
                 .feedDetailLoc(feedDetailLoc)
@@ -153,13 +179,26 @@ private final AwsS3Service awsS3Service;
     }
 
     @Transactional
-    public void modifyFeedComment(Long commentId, FeedRequestDto.FeedRequestCommentModifyDto feedRequestCommentModifyDto){
+    public void modifyFeedComment(User user, Long commentId, FeedRequestDto.FeedRequestCommentModifyDto feedRequestCommentModifyDto) {
+        // 댓글을 올린 사람만 권한이 있어야함
+        List<FeedComment> myFeedComment = commentRepository.findByIdAndUserId(commentId, user.getId());
+
+        if (myFeedComment.isEmpty()) {
+            throw new AuthFeedCommentNotFoundException();
+        }
         FeedComment feedComment = commentRepository.findById(commentId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
         feedComment.update(feedRequestCommentModifyDto);
 
     }
 
-    public void deleteFeedComment(Long commentId){
+    public void deleteFeedComment(User user, Long commentId) {
+        // 댓글을 올린 사람만 권한이 있어야함
+        List<FeedComment> myFeedComment = commentRepository.findByIdAndUserId(commentId, user.getId());
+
+        if (myFeedComment.isEmpty()) {
+            throw new AuthFeedCommentNotFoundException();
+        }
+
         commentRepository.deleteById(commentId);
     }
 }
